Reviewed by Adam Barth.
webkit-patch post, post-commits, upload should warn when posting to a closed bug, and offer to reopen it
https://bugs.webkit.org/show_bug.cgi?id=32006

I decided not to make it warn, and just have it re-open the bug.
That's not that different from today's behavior which will
just silently attach the patch.

This patch makes behavior between upload and land-safely consistent
(previously one would assign patches and the other would not)
as well as adds the ability for both to ensure that the bug is open.

To test this I had to add a few more methods to MockBugzilla which
(positively) affected a few other test results.

I also made AbstractStep keep a cached copy of the Bug object
and used the cached copy where appropriate (including for 'bug_title').
This should reduce the number of bug fetches we perform.

* Scripts/webkitpy/tool/commands/download_unittest.py:
* Scripts/webkitpy/tool/commands/upload.py:
* Scripts/webkitpy/tool/commands/upload_unittest.py:
* Scripts/webkitpy/tool/mocktool.py:
* Scripts/webkitpy/tool/steps/__init__.py:
* Scripts/webkitpy/tool/steps/abstractstep.py:
* Scripts/webkitpy/tool/steps/closebug.py:
* Scripts/webkitpy/tool/steps/ensurebugisopenandassigned.py: Added.
* Scripts/webkitpy/tool/steps/postdiff.py:
* Scripts/webkitpy/tool/steps/postdiffforcommit.py:
* Scripts/webkitpy/tool/steps/preparechangelog.py:
* Scripts/webkitpy/tool/steps/steps_unittest.py:
* Scripts/webkitpy/tool/steps/updatechangelogswithreviewer.py:

Identifier: 90305@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101834 268f45cc-cd09-0410-ab3c-d52691b4dbfc