[Gtk] Check for the load-complete event only in the unit tests.
There is currently a known issue happening in debug builds when
reloading the webview if it was loaded before through the function
webkit_web_view_load_string, resulting in a ASSERT failing (see
bug 50331) so we can't check for the reload event here, since it
makes the test to fail. Fortunately, checking for load-complete
event is a good enough test, so this small patch is about that.

Rubber-stamped by Philippe Normand.

* tests/testatk.c:
(testWebkitAtkDocumentLoadingEvents): Don't load and then reload
to check events, just load once and check whether the
'load-complete' event has happened.

Identifier: 90423@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101960 268f45cc-cd09-0410-ab3c-d52691b4dbfc