WebKit Mac does not build without CONTEXT MENU https://bugs.webkit.org/show_bug.cgi?id=73962
Patch by Benjamin Poulain <bpoulain@apple.com> on 2011-12-06
Reviewed by Pavel Feldman.

In the patch r100903, the symbols were exported under ENABLE(CONTEXT_MENUS)
because the feature is triggered from the menus.

The implementation has no dependency on the context menu but is necessary to build
when the inspector is enabled.
This patch moves the exported symbols from ENABLE(CONTEXT_MENUS) to ENABLE(INSPECTOR).

* WebCore.exp.in:


Identifier: 90638@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102205 268f45cc-cd09-0410-ab3c-d52691b4dbfc