WebPImageDecoder should not do a full image decode if progressive decoding is active https://bugs.webkit.org/show_bug.cgi?id=74041
Reviewed by Adam Barth.

If the decoder input data state reaches allDataReceived during a progressive image
decode, the decoder performs a full image decode.

On allDataReceived, check if we already have a decoder, and if so, continue to run
the progressive decoder.

No new tests. No change in behavior.

* platform/image-decoders/webp/WEBPImageDecoder.cpp:
(WebCore::WEBPImageDecoder::decode):


Identifier: 90725@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102310 268f45cc-cd09-0410-ab3c-d52691b4dbfc