CSS 2.1 failure: outline-color-* tests fail https://bugs.webkit.org/show_bug.cgi?id=71931
Source/WebCore:

Reviewed by Julien Chaffraix.

Test: css2.1/20110323/outline-color-001.html

WebKit didn't paint the top block in this series of tests because it ignored the outline
of objects with a zero size. Fix this by taking account of both offset and width of the
outline when deciding whether to paint it.

* rendering/RenderObject.cpp:
(WebCore::RenderObject::paintOutline): paint the outline even when the block has zero size

LayoutTests:

Add outline-color-001.htm to catch regressions against displaying outlines
on objects with zero size. The rest of the outline-color-* test series will
be landed separately.

Reviewed by Julien Chaffraix.

* css2.1/20110323/outline-color-001-expected.html: Added.
* css2.1/20110323/outline-color-001.htm: Added.

Identifier: 90924@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102526 268f45cc-cd09-0410-ab3c-d52691b4dbfc