[Refactoring] In prepare-ChangeLog, move into a method the top-level code to show ChangeLogs diff and open ChangeLogs in an editor https://bugs.webkit.org/show_bug.cgi?id=74266
Reviewed by Ryosuke Niwa.

We are planning to write unit-tests for prepare-ChangeLog
in a run-leaks_unittest manner. This patch is one of the incremental
refactorings to remove all top-level code and global variables from
prepare-ChangeLog.

* Scripts/prepare-ChangeLog: Moved the top-level code to show ChangeLogs diff into printDiff().
Moved the top-level code to open ChangeLogs in an editor into openChangeLogs().
Renamed @changed_files to @changedFiles.
Renamed %conflict_files to %conflictFiles.
(generateFunctionLists):
(printDiff):
(openChangeLogs):


Identifier: 90960@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102562 268f45cc-cd09-0410-ab3c-d52691b4dbfc