[V8] CodeGeneratorV8.pm does not correctly work with inherited HasIndexGetter https://bugs.webkit.org/show_bug.cgi?id=74027
Reviewed by Adam Barth.

Instead of having to write a custom indexer when the interface has an inherited indexer
we get the signature from the super interface.

Tested by existing tests.

* Target.pri: Remove V8DOMSettableTokenListCustom.cpp and V8WebKitCSSFilterValueCustom.cpp.
* UseV8.cmake: Ditto.
* WebCore.gypi: Ditto.
* bindings/scripts/CodeGenerator.pm:
(FindSuperMethod): Returns the first matching function in one of the ancestor interfaces.
* bindings/scripts/CodeGeneratorV8.pm:
(GenerateImplementationIndexer): If the current interface has no indexer try to find an
                                 indexer in one of the super interfaces.
(GenerateImplementationNamedPropertyGetter): Ditto for named property getter.
* bindings/v8/custom/V8DOMSettableTokenListCustom.cpp: Removed.
* bindings/v8/custom/V8WebKitCSSFilterValueCustom.cpp: Removed.


Identifier: 91038@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102646 268f45cc-cd09-0410-ab3c-d52691b4dbfc