[Refactoring] In prepare-ChangeLog, make several global variables be used only through parameter passing https://bugs.webkit.org/show_bug.cgi?id=74497
Reviewed by Ryosuke Niwa.

We are planning to write unit-tests for prepare-ChangeLog in a run-leaks_unittest
manner. This bug is one of the incremental refactorings to remove all top-level
code and global variables from prepare-ChangeLog. In this patch,
we make the following global variables be used only through parameter passing.
    - $bugDescription
    - $bugNumber
    - $name
    - $emailAddress
    - $gitReviewer
    - $writeChangeLogs

* Scripts/prepare-ChangeLog:
(fetchBugDescriptionFromURL):
(generateNewChangeLogs):
(reviewerAndDescriptionForGitCommit):


Identifier: 91352@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102988 268f45cc-cd09-0410-ab3c-d52691b4dbfc