[GTK] Make distcheck fails during the install https://bugs.webkit.org/show_bug.cgi?id=74274
No review, since this is a build fix.

* GNUmakefile.am: Remove BUILT_SOURCES from the dependency list for the gtkdoc
step. BUILT_SOURCES includes forwarding header generation for WebKit2, which
always runs. This means that the gtkdoc step was always running when make was
invoked. Generating gtkdoc during 'make install' was triggering a race condition
with the library file. Later we can fix generate-forwarding-headers and unbreak
'make docs,' but this bandaid is sufficient to let us release.

Identifier: 91402@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103038 268f45cc-cd09-0410-ab3c-d52691b4dbfc