Allow a PlatformCALayer to own its own sublayers https://bugs.webkit.org/show_bug.cgi?id=74744
Reviewed by Anders Carlsson.

GraphicsLayerCA rebuilds the sublayer list of CALayers, which would
blow away any custom layers that a PlatformCALayer wants to maintain
as children.

Make it possible for a PlatformLayerCA to indicate that it wants
a specific list of sublayers to be maintained as the first layers
in the child list.

* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::updateSublayerList):
* platform/graphics/ca/PlatformCALayer.h:
(WebCore::PlatformCALayer::customSublayers):
* platform/graphics/ca/mac/PlatformCALayerMac.mm:
(PlatformCALayer::PlatformCALayer):

Identifier: 91467@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103103 268f45cc-cd09-0410-ab3c-d52691b4dbfc