AX: aria-describedby should not be part of AXDescription (should be a part of AXHelp) https://bugs.webkit.org/show_bug.cgi?id=75052
Reviewed by Darin Adler.

Source/WebCore:

According to the ARIA spec, aria-describedby is a more detailed description that is akin to help, rather
than a description describing the actual object.

As such, it makes more sense for this attribute to be exposed under help text.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::helpText):
(WebCore::AccessibilityRenderObject::ariaAccessibilityDescription):

LayoutTests:

* accessibility/aria-describedby-on-input.html:
* platform/mac/accessibility/aria-describedby-on-input-expected.txt:
* LayoutTests/platform/gtk/Skipped
     Skip aria-describedby-on-input.html on GTK until helpText() can be implemented in DRT.



Identifier: 91860@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103548 268f45cc-cd09-0410-ab3c-d52691b4dbfc