Add a tool to analyze change logs https://bugs.webkit.org/show_bug.cgi?id=72243
Reviewed by Eric Seidel.

Add the initial implementation of analyze-changelog command. This version supports parsing multiple changelogs
in a single directory and summarizes the result in json files. It also copies summary.html to the output directory
for visualizing json files.

Usage: webkit-patch analyze-changelog [--changelog-count N] <CHANGELOG_DIR> <OUTPUT_DIR>
where CHANGELOG_DIR is the directory to look for changelogs, OUTPUT_DIR is the directory into which the tool
generates json files, and N is the number of changelogs to parse.

* Scripts/webkitpy/common/config/contributionareas.py: Added names().
* Scripts/webkitpy/common/system/filesystem_mock.py: Added readline and iterator support to ReadableTextFileObject.
* Scripts/webkitpy/tool/commands/__init__.py:
* Scripts/webkitpy/tool/commands/analyzechangelog.py: Added.
* Scripts/webkitpy/tool/commands/analyzechangelog_unittest.py: Added.
* Scripts/webkitpy/tool/commands/data/summary.html: Added.
* Scripts/webkitpy/tool/steps/options.py:


Identifier: 92256@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103959 268f45cc-cd09-0410-ab3c-d52691b4dbfc